+++ src/containers/MainContainer.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { BrowserRouter as Router, HashRouter, Route, Routes } from 'react-router-dom'
 
-import { LoginForm } from '../components/formComponents/LoginForm';
-import { RegisterForm } from '../components/formComponents/RegisterForm';
+import { LoginForm } from '../components/oldFormComponents/LoginForm';
+import { RegisterForm } from '../components/oldFormComponents/RegisterForm';
 import { MainDisplay } from './MainDisplay';
